+++ lib/instrumentation/express.js
@@ -127,7 +127,7 @@ function wrapResponse(shim, response) {
         if (!shim.isFunction(cb)) {
           ++cbIdx
           cb = function defaultRenderCB(err, str) {
-            // https://github.com/expressjs/express/blob/4.x/lib/response.js#L961-L962
+            // https://github.com/expressjs/express/blob/4.x/lib/response.js#L961-L962
             if (err) {
               return res.req.next(err)
             }
