+++ test/versioned/generic-pool/package.json
@@ -12,9 +12,8 @@
         "generic-pool": ">=3.0.0"
       },
       "files": [
-        "basic.tap.js"
+        "basic.test.js"
       ]
     }
-  ],
-  "dependencies": {}
+  ]
 }
