+++ test/versioned/grpc/package.json
@@ -12,14 +12,14 @@
         "@grpc/grpc-js": ">=1.4.0"
       },
       "files": [
-        "client-unary.tap.js",
-        "client-streaming.tap.js",
-        "client-server-streaming.tap.js",
-        "client-bidi-streaming.tap.js",
-        "server-unary.tap.js",
-        "server-client-streaming.tap.js",
-        "server-streaming.tap.js",
-        "server-bidi-streaming.tap.js"
+        "client-unary.test.js",
+        "client-streaming.test.js",
+        "client-server-streaming.test.js",
+        "client-bidi-streaming.test.js",
+        "server-unary.test.js",
+        "server-client-streaming.test.js",
+        "server-streaming.test.js",
+        "server-bidi-streaming.test.js"
       ]
     }
   ]
